Update Geocodable to use Decimal rather than Float

+++ code/Geocodable.php
@@ -8,8 +8,8 @@
 class Geocodable extends DataExtension {
 
 	public static $db = array(
-		'Lat' => 'Float',
-		'Lng' => 'Float'
+		'Lat' => 'Decimal',
+		'Lng' => 'Decimal'
 	);
 
 	public function onBeforeWrite() {
